Changed Sample.update() to `intersect' distributions rather than `uniting' them; i.e. multiply pointwise, rather than adding.  Did this by adding a .cross method to joinWeighted, supported in turn by interpolator.cross (including a bodge to cope with combining incompatible distributions).
Ripped interpolator out of repWeighted into a new base-class,
_curveWeighted (along with the entropy-related toys that didn't belong
in repWeighted), so that joinWeighted and repWeighted can both inherit
from it.  May make replacement interpolators easier to do.

Also, changed Sample.update() to only mess with distribution if offered
new value is also a distribution, otherwise only add it to a list of
`best' estimates whose median to use as actual .best attribute; this
list also gets an entry from any distributions offered, if they're
packaged as Sample() so have a .best to offer.  Stopped bothering with
(hopefully) unused weight/func args specific to add(), which was the
crux of the old implementation.  Skeptical about retaining **what.