Re-worked mutual to handle the case of a pair of attributes.
Extended the reflexive-attribute class, lazily based on each class
derived from weakprop, to deal with the case of a pair of attributes,
each of which refers back via the other, as x.a.b is x is
x.b.a. Documented why the generalization to longer cycles is messy.
Killed CVS $Id$ tag, now that we're in git.