Broke out most of __back()'s class's code into a shared base.
Define a base-class for the mutual-attribute properties, as as to limit
duplication of code among the derived classes created by __back().
This also helps localise use of weakref slightly better; and provides a
better home for the fat comment about longer cycles of properties.
Fixed assorted infelicities in the process.