Implement keeping the SVG element itself tidy.
Pursuant to limiting commit diffs to what's actually changed, with as
little extra content on the changed lines as practical: I want the
viewBox on a line of its own; and I reflow the other attributes to avoid
overly long lines, which I'd like the script to also preserve.

Implemented parsing of the file's emacs mode-line so that we respect
fill-column and tab-width.  Killed CVS $Id$.