Made joinWeighted.__combine more faithful.
As its documentation indicated, its choice of where to put weights was
distinctly suspect.  Implemented the most obviously less wrong choice.

This made the interpolator's .cross()'s rather funky use of .combine()
invalid (which, frankly, it was anyway), requiring minor re-working.

Since the new approach wants function values at interval*interval
corners, it's now necessary to cope with evaluating functions at
boundaries of intervals, which are apt to be zero, which creates
problems when dividing.  So the new solution includes some kludging to
handle that, with extensive documentation; see corners().