Broke out the interpolator from sample.py
This begins a long-over-due clean-up; curveWeighted's .interpolator had
better properties than weighted-value dictionaries, so it's time to
replace the latter with systematic use of interval-based distributions,
such as the interpolator.

First step is to break out the interpolator as a class in its own right,
in study.maths.interpolator; base-class Interpolator defines the API and
interpolator-independent basics, while PiecewiseConstant provides the
implementation details specific to the piecewise-constant model I've
been using.  Interaction with weighted-value dictionaries is now handled
by curveWeighted.Interpolator, which is used to implement
_lazy_get_interpolator_(), now as a method.

Had to move Atmosphere.technical from units.py to archaea, as its use of
.weight (is non-SI-ish and) required the kind-specific attribute
machinery for masses, one (of the others) of which needs an import from
study.chemy.physics, which imports units.