Use interpolators more pervasively in value/sample.py
Constructors for weighted-value dictionaries now accept an interpolator
as well as or instead of the dictionary; and all internal calls between
weighted-value classes propagate interpolators.  Next step, when I can
find the time, is to push the dictionaries out to the Sample level and
convert to work exclusively with interpolators internally.