Accumulated minor details.
Lots of CVS $Id$ tags that were in a different format to the one in
force when I imported to git.  Several reflows.  A few comments
(including a big slab of data in element.py) and documentation tweaks.
A stray import suppressed in an __init__.  Other trivia.
A not-yet-working Ocean.volume property.