Moved combination machinery from weighted to interpolator.
Specifically, moved implementation details of joinWeighted.combine() to
new Interpolator.combine() method with implementation in
PiecewiseLinear.  At the same time, did the pending TODO and converted
to dealing with each pair of intervals as a rectangle in co-ordinate
space mapped to a suitable piecewise linear distribution in output
space, to be combined with others to produce a suitable composite.

Fixed flawed handling of divide-by-zero - the implementation didn't
match the theory; and the theory had room for improvement, too !

Chose .combine(func, other) as signature to leave room for making it
.combine(func, *others) some day.

Added Interpolator.simplify(), using n-iles in the obvious way.

Assorted bug-fixes and clean-up.
In PiecewiseConstant:
 * .split() could get upset by empty late bands; taught it how to calmly
   fill them with zeros.
 * .weigh()'s handling of boundary spikes wasn't quite right; it gave
   them wrong weights and could spin forever; taught it how to do the
   job right in finite time.
In sample.py:
 * killed _baseWeighted's bogus .low() and .high().
 * taught joinWeighted.add() to cope gracefully when there's too little
   data to build an interpolator.
 * made joinWeighted.condense() more insistent about getting the right
   outer bounds
 * scrap clunky old mean/variance computation in favour of
   Interpolator.normal
 * less overt handling of interpolator in (pseudo-) constructors
 * use cls, not klaz, for name of variable that's a class
Generally:
 * detect and raise various errors
 * change some float comparisons to "tiny difference" checks
 * assorted simplification
 * relevant comments
 * del stray tools after use
 * typo fixes