Substantial over-haul of Vector class.
Reflowed lots of documentation, broke up some long lines.  Provide
Tensor as an alias for Vector and pseudo-constructor .fromSeq to recurse
into a sequence [of sequences ...] to make all non-leaf entries Vector
objects.  Define __str__ to not waste space over the class name and
__repr__ to only do so at outermost level, exploiting .fromSeq() to keep
it valid.

Added .dot(), .rdot() as simple contraction.
Added pseudo-constructors .xerox, .delta and .diagonal for really easy
cases; .circulate() and .hyperbolate() for particular common cases.
Added .embed() to make it easy to do those in higher dimension.
Fixed a bug in transpose: 0 is natural !

Did much of this a few weeks back, the rest today.