Added Fibpair and refined existing parts.
Fibpair implements the underlying modular polynomials used by fastonacci
and provides more flexibility.  Added iterator support and partial "is
in" support for the cached Fibonacci sequences (and renamed a local
variable).  Fixed an out-of-date reference to BigFloat.  Reworked fibpow
to cope with negative n and take the right arguments to make fastonacci
simple.