Some tidy-up in Quantity, preparing for coming changes.
Break out tonumber() tool-function from Prodict.__pow__ so that I can
re-use it.  Quantity constructor doesn't need to coerce units to Bok
when decomposing scale as a Quantity worked, so move the coercion to the
exception clause.  The similar code in .flat() multiplied the two units
values the suboptimal way round, so swapped them.  Used .__kin() a bit
more widely.   Commented on a deprecation and a TODO.