More pseudo-constructors for Quantity; and tidied .flat()
Nothing passed .flat() a name or sample, so drop support for those being
explicit: they can be passed as keywords anyway.  Support a rescale
parameter that has the same effect as unit, when it's a Quantity, but
can be a simple number (i.e. needn't be a Quantity), which unit can't
be.

Added .within(), .below(), .fromSpread(), .fromDecimal(),
.fromSigFigs(), and .gaussian() pseudo-constructors.  Deprecated the
rest of the particular Quantity objects in the tail, along with
sample(), in favour of using these methods, and .flat(), to do the jobs
they supported.