Clean-up in radix050.py
It used repeated doubling while catching OverflowError, which won't
happen any more (doubling a big int simply turns it into a long), to
find maxint.  Import it from sys and clean up the code that used it, now
that I'm on a 64-bit box and it no longer makes sense to just assume int
is a 32-bit type.  Provided for decode to be told the wordsize actually
relevant when kludging negative values into usability.