Lots of petty tidy-up in python code.
Purged CVS cruft.  Use cls as name for class-parameters (and local
variables).  Prefer isinstance() checks over comparison of type()
values.  Stop worrying about int->long OverflowError; it hasn't happened
in ages.  Use True and False as booleans in preference to 1 and
None.  Tidied some code as a result of all of that.  Converted tabs to
spaces in some cases.

Fixed a few trivia that I picked up along the way.
Element._lazy_get_atom_ computed a value for A but forgot to store it.
A few things referred to maths.permute.Permute which is actually called
Permutation.