Use any, all and comprehensions in place of filter and map.
There's plenty more of these to replace, but these were the ones it was
easy to find | xargs grep for - and particularly clear examples of the
new way being more readable, too.  In the process, fixed interpolator's
constructor to actually work if given iterators instead of sequences, as
advertised.  Fixed a few comparisons to initial substrings to use
.startswith(), while I was about it.