Moved definition of astronomical unit from space.home to value.units
The IAU has lately revised the definition to a whole number of
hectometres, making it more apt to define it among the SI-compatible
units.  It also gets a canonical (lower-case) name, au, so it gets to be
study.value.units.au, although study.space.home still exports all its
old names, as well as the new.