More documentation, options and other refinements for voronoi.Catchment
Replaced the same=True parameter of .nearest() with an elide=()
parameter, for greater generality (at no cost in code complexity; and
it's easier to explain in the doc).  Fixed handling of this (when same
was True, so that ancestors of the box containing where could be empty
if their other leaves are, one of these ancestors could cause the actual
nearest neighbour's box to be eliminated in .__nearleaves()) by setting
.points on all nodes in the BTree, not just the leaves.

Reworked BTree somewhat.  Splitting now prefers to happen on a less
recently used index even if a more recently used one has a smaller
diameter (to avoid bad behaviour if many points lie on some hyperplane
with one coordinate constant).  Added a .parent attribute in support of
that.  Details have shifted around a bit among BTree's private methods
in the process, notably turning its @classmethod .__btree() into a
.__child() instance method.

Made the internal bisection tree private (s/\.tree/.__root/wg).
Let client tune the number of points in each BTree.
Documented the classes and most methods.
Added page-splits in strategic places.