Make Vector more amenable to sub-classing.
Replaced fromSeq()'s dimension parameter (which was only used in two
places, in both of which it was used fatuously) with a template from
which to infer which classes to use at each step of collecting values up
into vectors.  Made __repr__ use the name of the actual class in use,
not just assume it hasn't been subclassed.