Fixed fiddly conditional to the right way round.
In voronoi.Catchment.__neareleaves()'s check for half of a split being
the new best node, I'd managed to confuse myself by having the if branch
be the "not as good" while the preceding comment's explanation related
to the "at least as good" condition.  So reversed the conditional to
make it take the form of the condition explained - and correct its
conditioning on whether the split node used to be the best.