Support negation; and some fixes in Namely.
Added Vector.__neg__().  Namely's check against None entries ran afoul of "is
in" tests using equality and somethings not liking to be compared vs None; so
fix to an any(x is None for x in seq) test instead.  The Namely.__vector__()
wanted to * its parameter, which is sometimes an iterator; so tuple() it first.