Complicated seismic details, added archaea.Krakatoa
Changed energies' .seismic to a sub-object with attributes matching those of
archaea.Magnitude; and recoded the latter (as a class) so that one can
round-trip between the two (which requires going via .evaluate() when
Magnitude's methods are called on a Quantity).

Added Krakatoa as a unit of energy on such a scale as to have climactic
impact.  Also corrected my (apparently) mistaken supposition that megatons use
the US ton; Wikipedia assures me it's the metric tonne.