Document Iterable.reduce() and handle init better.
Also @lazyprop mean and geomean, like max and min, on ReadOnlySeq.  For sum and
product, use reduce()'s default for init, so as to avoid having to pass them a
suitable additive identity when it shouldn't be needed, instead of supplying
defaults of our own; but still let the caller pass init if wanted.  Make mean()
fussy about preserving wholeness, when it can.