More iterator-friendly version; and improved Dict.cartesian()'s doc.
Make less of the old functional builtins, in favour of list comprehensions and
iterators.  Tweaked @iterable to be able to apply to a static method with no
args.  Renamed Dict.cartesian()'s arglist to make the doc-string read better -
once I'd fixed some infelicities in it (no longer an instance method, so no self
to prepend to the *others) - and gave an example of using this method.