A helper tool for use with valgrind's memcheck logs.
Can parse log files, collecting reporrts of leaks and other issues.  Can iterate
over these, skipping ones marked as fixed; I can tell it when I've repaired an
error and have it mark all issues, whose stack includes a specific frame, as
fixed.  It also propagates that in some cases: when the last reader of an
uninitialized slab of memory, or last accessor of a freed slab, has been
resolved, the badly-accessed memory issue also clears itself.  This makes it
easier to keep track of which stacks still need to be fixed, as compared to
ignoring stack traces that look like ones I think I've fixed before.  There is
significant scope for further improvement here, but this is a useable first
pass.