Split iterable from iterinstance.
The magic treatment of the decorated function's first argument made it
impossible to control whether a class based on Iterator got used as the wrapper
or not.  When Permutation (based, via Tuple, on Iterable) wants an iterator that
isn't a permutation, this got problematic.  So split the decorator into two and
oblige the author to be clear about whether to make an instance of the present
class or just make a simple iterator.