Enhance Module to package imports within modules it loads.
Previously, when the source file for a directory did an import of other python
modules within the directory, or of a sub-directory, they'd be represented by
ordinary module objects, without auto-loading functionality.  So rework the
loading to massage any module objects loaded into a namespace to be represented
via the object our auto-loading would have loaded in its place !  This is all
just one huge kludge-around for now knowing a way to offer the import machinery
a callable of my chosing that it can use in place of the builtin module().

Also gave the root object a smarter default name, tweaked the repr and docs.