Exctract a factor() from isprime() and tidy up a bit in natural.py
The implementation of isprime() could actually returna proper factor (when input
isn't 0 or a unit), so break that function out of it.  Revised a slight
imprecision in the proof that a rational is whole if its square is.  Refined
documentation of desquare, turned one comment into an assertion, added another
comment.