Play correctly with the data descriptor getter protocol.
While the getter functions for properties don't need a cls=None parameter, the
__get__ of a data descriptor does, albeit I'm not using it, as the type of the
object gets passed as this parameter.  (It's not clear when, if ever, __get__ is
called *without* this parameter, so the =None may be redundant, but the howto
guide seems to believe it's needed.)  When used to get the property of the
class, rather than of an instance, obj is None (which has no .__dict__, so
dictprop.__get__() catches that) and the property itself is the thing to return.