Have Polynomial accept an iterator over (power, coefficient) twoples.
This requires that the constructor traverses the data only once, so move the
elimination of common factors to after we've got it all safely in .__coefs; we
can rewrite it once we've found what to eliminate.  The check that powers are
whole also has to move into .__store() for the iterator to hit it.

Add class methods .fromSeq() and .fromMap() to take over the old ways of
supplying coefficients; the constructor still supports those, but reliance on
that is now deprecated.  Gave .power() optional scale and denominator parmeters
to make it more useful; added a .copy() method.  Converted all clients of the
class to use the new API.

Redirected all instantiation via a pseudo-constructor ._polynomial_() so that
derived classes have the option of over-riding, but kept it using raw Polynomial
since no extant derived class wants that (and this matches what we had before).
That turns various static methods into class methods.

Cut out some uses of ._zero; fixed a glaring bug in .integral() and made it use
the denominator facility rather that resort to float divisions - now with
parentheses round the 1 + k denominator !