Improvements to MHTML digestion.
If a Content-Location is an URL, use its path, discarding scheme,
host, query string and fragment identifier, if present.  Use .html or
.xhtml as suffix for HTML files, not (e.g.) .asp

Avoid infinite loop in MHT.stem by exiting early on empty; and use
rfind() to save a hand-coded implementation of it !

Avoid infinite recursion in __mkdirs() caused by os.path.spli('/')
being ('/', ''); if the head of the path is the whole path, give up.