Refactor pythagorean triangle iteration.
Provide a plain iterator that just iterates all pythagorean triangles,
filtering with a test, which default to checking coprime.  This makes
it possible to also select triangles approximating some particular
ratio of sides, for example.  Refactor the old iterator to use the
plain one.  (Not really sure it's even worth keeping; all it does is a
bunch of complicated cacheing, of debatable value.)