Record Sun's magnetic moment and tidy Earth's.
These can now be inferred from surface field strength, so move parts
around to make the most of that.  Also standardise on magnetic as
intermediate Object, with field as field strength or dipole as total
dipole (being made into a lazy attribute of the Planet class), and add
a comment on how spin and Moon.orbit.radius have varied.