Fix various mistakes in maths.norm
Adjusting scale, when it was supplied by the caller, could have
overflowed, in one class of cases, so deal with that.

It turns out cmath has no hypot(), so don't try to delegate two-arg
hypot() to it; and math.hypot() won't do either, as it can't cope with
complex args.

The L1-norm should be infinity if any entry is infinite, even if some
entries are NaN, so don't simply sum abs.