Tweak Job's defaults and doc.
In Norway, we have six holidays per year tied to Easter, hence on
specific days of the week and definitely not at week-ends, plus four
and three halves that float around the week, so only happen five years
in seven on average; and my employer lets me have the three halves as
full days, for an average of 7 * 5 / 7 = 5; with the six on fixed days
of the week, that makes 11; and I work 7.5 hour days, five days a
week, with 25 days off (plus any I carried over from an earlier year).

(This at-work change seems to overlap an at-home change that adjusted
the number of public holidays already. Combined on rebasing.)