Evaluate integrand() inside interval when clipping.
When the integrand is evaluated outside the range in which the
Integrator is defined to operate, we need to evaluate the underlying
function to get a value of the right kind, that we scale by zero to
return a zero of the right kind.  However, the value given being
outside the range, it's possible the underlying function is
ill-behaved there, so select a value inside the range at which to
evaluate it.  Use the end-point of __span() that isn't the value
passed, since that's implicitly in the function's range.