Use a more generous width to blur a value
When blurring the final result of an integration, use the abs() of the
last step and, when available, the width of the total + offset we
compared it to when deciding not to bother with further refinement,
rather than only the (possibly negative) last step.  Likewise, for the
default offfset, take the abs of total's difference from its own best
and use total's width if available and bigger.