Rework thermal.Radiator to handle quantities using units of angle.
Frequency, wavelength and their friends intrinsically contain units of
angle; and orthodoxy contains traditions that implicitly use turn and
radian as the unit for that.  Instead of having distinc parameters for
'angular' (i.e. radian-based) frequency and (tacitly cycle-based)
frequency, actualy include units of angle in the quantities.  Retain
support for these values without units of angle, though, presuming
turn as implicit unit of angle for such values.

Rework spectral() to take any of the values that identify a point in
the spectrum, detecting which it's been given and converting
accordingly, both to get a frequency from the input and to turn the
radiance per frequency result into a radiance per change in input.

Add a pseudo-constructor fromPeak() that takes any of the same
quantities, constructing an instance whose peak radiance with respect
to that quantity is the value given.  Expose temperature as a
(read-only) property, so that you can look it up on such an instance.
Turn splitexp() into a class variable __splitexp that stores the two
values of it we actually need, now that these are each used in two
places.