PowerTails.__split(): rework error estimation.
The error in a strip is (an eigth of) the change in gradient times the
square of interval width; dividing this by the strip area gives the
change in greadient times the interval width divided by the density.
Since it's relative error we want to bound, we no longer need __scale,
simplifying __cuts().  The computation is now acceptably swift even
with a fractional error constraint of 1e-6, so change tiny's default.